refactor(parser): split stmt/simple_expr_stmt.rs into cohesive submodules (§7-8)
Extract the 2134-line monolith into a facade + 5 cohesive submodules:
- core.rs: expr_stmt() function (primary statement parser)
- let_temp.rs: let_stmt() and temp_stmt()
- lvalue.rs: lvalue builders and bind helpers + decl_target_var_name
- predicates.rs: term-detection and pure-value predicates
- sig_info.rs: SigParamInfo + signature binding extraction helpers

All super:: paths rewritten to absolute crate:: paths. fn/type
multiset byte-identical — pure relocation, no logic changes.
